add multiple new fetures
almost finalized the song layout and how it looks, just need to add the links to other songs that are related to it like songs in the same album. Also, add an updating top 10 list for the songs page with hyperlinks with animation to the list when opening the page.